added Tree2 example (issues)
Again we have the Change widget problem. In this case (and also Tree1)
though it seems this is confusion between behaviour when a symbol (:usr)
etc. is used and when we pass text to a Term. Consistent use of Id(s)
solves the propblem (however this can surely introduce some subtle
errors)